feat: Add tts tooltip
