--- src/common/components/Modal/modal.module.scss
@@ -1,13 +1,0 @@
-.modal {
-  animation: animate_modal 1s ease-in-out forwards;
-}
-
-@keyframes animate_modal {
-  from {
-    opacity: 0;
-  }
-
-  to {
-    opacity: 1;
-  }
-}
